Initial refactoring of the coordinator selection code
